Leetcode/Sort/BubbleSort.cs
using System;
using System.Collections.Generic;
using System.Text;

namespace Leetcode.Sort
{
    public class BubbleSort
    {
        public static void Sort(int[] numbers)
        {
            for (int i = 0; i < numbers.Length; i++)
            {
                for (int j = 0; j < numbers.Length; j++)
                {
                    if (numbers[i] < numbers[j])
                    {
                        var temp = numbers[i];
                        numbers[i] = numbers[j];
                        numbers[j] = temp;
                    }
                }
            }
        }
    }
}
